requirements.txt
opencv-python
numpy
pyserial
python-dotenv
pyautogui
